# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "FairnessIntake": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "FairnessIntake",
        "type": "object"
      },
      "FairnessOut": {
        "properties": {
          "intake": {
            "$ref": "#/components/schemas/FairnessIntake"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "stages": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/GroupRates"
              },
              "type": "object"
            },
            "title": "Stages",
            "type": "object"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "intake",
          "stages",
          "warnings",
          "note"
        ],
        "title": "FairnessOut",
        "type": "object"
      },
...
      "FunnelIntake": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quota": {
            "title": "Quota",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "quota",
          "status"
        ],
        "title": "FunnelIntake",
        "type": "object"
      },
      "FunnelOut": {
        "properties": {
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Status",
            "type": "object"
          },
          "decisions": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Decisions",
            "type": "object"
          },
          "intake": {
            "$ref": "#/components/schemas/FunnelIntake"
          },
          "median_days_to_decision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Days To Decision"
          },
          "quota_fill": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Fill"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/FunnelStage"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "intake",
          "by_status",
          "stages",
          "decisions",
          "median_days_to_decision",
          "quota_fill"
        ],
        "title": "FunnelOut",
        "type": "object"
      },
      "FunnelStage": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "count"
        ],
        "title": "FunnelStage",
        "type": "object"
      },
      "GroupRates": {
        "properties": {
          "impact_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Ratio"
          },
          "rates": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Rates",
            "type": "object"
          },
          "sizes": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Sizes",
            "type": "object"
          }
        },
        "required": [
          "rates",
          "sizes",
          "impact_ratio"
        ],
        "title": "GroupRates",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabAnalysis": {
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/LabCriterionStat"
            },
            "title": "Criteria",
            "type": "array"
          },
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "model_auc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Auc"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "qualified_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualified Rate"
          },
          "reliable": {
            "title": "Reliable",
            "type": "boolean"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "n",
          "events",
          "criteria",
          "model_auc",
          "warnings",
          "reliable"
        ],
        "title": "LabAnalysis",
        "type": "object"
      },
      "LabCriterionDef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "weight"
        ],
        "title": "LabCriterionDef",
        "type": "object"
      },
      "LabCriterionStat": {
        "properties": {
          "auc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auc"
          },
          "ci90": {
            "items": {
              "type": "number"
            },
            "title": "Ci90",
            "type": "array"
          },
          "coef": {
            "title": "Coef",
            "type": "number"
          },
          "criterion": {
            "title": "Criterion",
            "type": "string"
          },
          "mean_if_not": {
            "title": "Mean If Not",
            "type": "number"
          },
          "mean_if_qualified": {
            "title": "Mean If Qualified",
            "type": "number"
          },
          "significant": {
            "title": "Significant",
            "type": "boolean"
          }
        },
        "required": [
          "criterion",
          "coef",
          "ci90",
          "significant",
          "auc",
          "mean_if_qualified",
          "mean_if_not"
        ],
        "title": "LabCriterionStat",
        "type": "object"
      },
      "LabIn": {
        "properties": {
          "intake_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Intake Ids",
            "type": "array"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "required": [
          "intake_ids"
        ],
        "title": "LabIn",
        "type": "object"
      },
      "LabOut": {
        "properties": {
          "admitted": {
            "title": "Admitted",
            "type": "integer"
          },
          "analysis": {
            "$ref": "#/components/schemas/LabAnalysis"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/LabCriterionDef"
            },
            "title": "Criteria",
            "type": "array"
          },
          "new_weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Weights"
          },
          "old_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Old Weights",
            "type": "object"
          },
          "pool": {
            "title": "Pool",
            "type": "integer"
          },
          "simulation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LabSimulation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "criteria",
          "analysis",
          "pool",
          "admitted",
          "old_weights"
        ],
        "title": "LabOut",
        "type": "object"
      },
      "LabSimulation": {
        "properties": {
          "changed_in": {
            "title": "Changed In",
            "type": "integer"
          },
          "dropped_ids": {
            "items": {
              "type": "string"
            },
            "title": "Dropped Ids",
            "type": "array"
          },
          "fairness_new": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GroupRates"
            },
            "title": "Fairness New",
            "type": "object"
          },
          "fairness_old": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GroupRates"
            },
            "title": "Fairness Old",
            "type": "object"
          },
          "model_expected_new": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Expected New"
          },
          "model_expected_old": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Expected Old"
          },
          "model_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Note"
          },
          "newly_selected_ids": {
            "items": {
              "type": "string"
            },
            "title": "Newly Selected Ids",
            "type": "array"
          },
          "observed_new": {
            "$ref": "#/components/schemas/ObservedRate"
          },
          "observed_old": {
            "$ref": "#/components/schemas/ObservedRate"
          },
          "overlap_actual_old": {
            "title": "Overlap Actual Old",
            "type": "number"
          },
          "overlap_old_new": {
            "title": "Overlap Old New",
            "type": "number"
          },
          "pool": {
            "title": "Pool",
            "type": "integer"
          },
          "selected": {
            "title": "Selected",
            "type": "integer"
          }
        },
        "required": [
          "pool",
          "selected",
          "overlap_old_new",
          "changed_in",
          "overlap_actual_old",
          "observed_old",
          "observed_new",
          "fairness_old",
          "fairness_new",
          "newly_selected_ids",
          "dropped_ids"
        ],
        "title": "LabSimulation",
        "type": "object"
      },
...
      "ObservedRate": {
        "properties": {
          "known": {
            "title": "Known",
            "type": "integer"
          },
          "of": {
            "title": "Of",
            "type": "integer"
          },
          "qualified_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualified Rate"
          }
        },
        "required": [
          "known",
          "of",
          "qualified_rate"
        ],
        "title": "ObservedRate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Talent Hub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analytics/fairness": {
      "get": {
        "operationId": "fairness_api_v1_analytics_fairness_get",
        "parameters": [
          {
            "in": "query",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FairnessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fairness",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/funnel": {
      "get": {
        "operationId": "funnel_api_v1_analytics_funnel_get",
        "parameters": [
          {
            "in": "query",
            "name": "intake_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Intake Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Funnel",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/lab": {
      "post": {
        "operationId": "lab_api_v1_analytics_lab_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lab",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}